Conflicts:
	debian/changelog

+++ etc/initsystem/icinga2.init.d.cmake
@@ -21,14 +21,17 @@ if [ -f $SYSCONFIGFILE ]; then
 	. $SYSCONFIGFILE
 else
 	echo "Can't load system specific defines from $SYSCONFIGFILE."
-	exit 1
+	exit 6
 fi
 
-test -x $DAEMON || exit 0
+test -x $DAEMON || exit 5
+
+ICINGA2_USER=`$DAEMON variable get --current RunAsUser`
+ICINGA2_GROUP=`$DAEMON variable get --current RunAsGroup`
 
 if [ ! -e $ICINGA2_CONFIG_FILE ]; then
         echo "Config file '$ICINGA2_CONFIG_FILE' does not exist."
-        exit 1
+        exit 6
 fi
 
 # Get function from functions library
@@ -48,7 +51,7 @@ start() {
 	printf "Starting Icinga 2: "
 	@CMAKE_INSTALL_FULL_SBINDIR@/icinga2-prepare-dirs $SYSCONFIGFILE
 
-	if ! $DAEMON -c $ICINGA2_CONFIG_FILE -d -e $ICINGA2_ERROR_LOG -u $ICINGA2_USER -g $ICINGA2_GROUP > $ICINGA2_STARTUP_LOG 2>&1; then
+	if ! $DAEMON daemon -c $ICINGA2_CONFIG_FILE -d -e $ICINGA2_ERROR_LOG > $ICINGA2_STARTUP_LOG 2>&1; then
 		echo "Error starting Icinga. Check '$ICINGA2_STARTUP_LOG' for details."
 		exit 1
 	else
@@ -65,7 +68,7 @@ stop() {
                 if [ "x$1" = "xnofail" ]; then
 			return
 		else
-			exit 1
+			exit 7
 		fi
         fi
 
@@ -94,12 +97,16 @@ stop() {
 reload() {
 	printf "Reloading Icinga 2: "
 
+	if [ ! -e $ICINGA2_PID_FILE ]; then
+		exit 7
+	fi
+
 	pid=`cat $ICINGA2_PID_FILE`
 	if kill -HUP $pid >/dev/null 2>&1; then
 		echo "Done"
 	else
 		echo "Error: Icinga not running"
-		exit 3
+		exit 7
 	fi
 }
 
@@ -107,7 +114,7 @@ reload() {
 checkconfig() {
 	printf "Checking configuration: "
 
-	if ! $DAEMON -c $ICINGA2_CONFIG_FILE -C -u $ICINGA2_USER -g $ICINGA2_GROUP > $ICINGA2_STARTUP_LOG 2>&1; then
+	if ! $DAEMON daemon -c $ICINGA2_CONFIG_FILE -C > $ICINGA2_STARTUP_LOG 2>&1; then
                 if [ "x$1" = "x" ]; then
 			cat $ICINGA2_STARTUP_LOG
 			echo "Icinga 2 detected configuration errors. Check '$ICINGA2_STARTUP_LOG' for details."
@@ -131,12 +138,17 @@ checkconfig() {
 status() {
 	printf "Icinga 2 status: "
 
+	if [ ! -e $ICINGA2_PID_FILE ]; then
+		echo "Not running"
+		exit 7
+	fi
+
 	pid=`cat $ICINGA2_PID_FILE`
 	if kill -CHLD $pid >/dev/null 2>&1; then
 		echo "Running"
 	else
 		echo "Not running"
-		exit 3
+		exit 7
 	fi
 }
 
@@ -165,6 +177,6 @@ case "$1" in
 	;;
   *)
         echo "Usage: $0 {start|stop|restart|reload|checkconfig|status}"
-        exit 1
+        exit 3
 esac
 exit 0
